% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chashmap.R
\name{get}
\alias{get}
\title{retreive values for specific keys from a hashmap}
\usage{
get(map, keys)
}
\arguments{
\item{map}{Hashmap object of class \code{C_hashmap}}

\item{keys}{a scalar or vector of type numeric (int or real) or strings}
}
\description{
retreive values for specific keys from a hashmap
}
