% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chashmap.R
\name{insert}
\alias{insert}
\title{insert elements into hashmap}
\usage{
insert(map, keys, values)
}
\arguments{
\item{map}{Hashmap object of class \code{C_hashmap}}

\item{keys}{a scalar or vector of type numeric (int or real) or strings}

\item{values}{a scalar or vector of type numeric (int or real) or strings}
}
\description{
insert elements into a hashmap
}
\details{
When inserting key value pairs as vectors, make sure that length(key) == length(value).
The insertion is vectorized over the vector. Ie. internally, the map will call \code{insert(map, key[i], value[i])}
for i in 1:length(key)
}
