% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chashmap.R
\name{keys}
\alias{keys}
\title{retreive keys from hashmap}
\usage{
keys(map, simplify = FALSE)
}
\arguments{
\item{map}{Hashmap object of class \code{C_hashmap}}

\item{simplify}{TRUE coerces the key list to a vector (default = FALSE)}
}
\description{
Retreive keys from the hashmap as a list
}
