% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chashmap.R
\name{remove}
\alias{remove}
\title{remove keys from hashmap}
\usage{
remove(map, keys)
}
\arguments{
\item{keys}{a scalar or vector of type numeric (int or real) or strings}
}
\description{
Remove keys from the hashmap
}
