% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chashmap.R
\name{size}
\alias{size}
\title{Return the number of elements stored in the hashmap}
\usage{
size(map)
}
\arguments{
\item{map}{Hashmap object of class \code{C_hashmap}}
}
\description{
Return the number of elements stored in the hashmap
}
