% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chashmap.R
\name{values}
\alias{values}
\title{Retreive values from hashmap}
\usage{
values(map, simplify = FALSE)
}
\arguments{
\item{map}{Hashmap object of class \code{C_hashmap}}

\item{simplify}{TRUE coerces the key list to a vector (default = FALSE)}
}
\description{
Retreive values from the hashmap as a list
}
